Refactor cache clearing and implement it
