+++ .pylintrc
@@ -83,7 +83,7 @@ ignored-argument-names = _.*
 max-locals = 15
 max-returns = 6
 max-branches = 12
-max-statements = 50
+max-statements = 55
 max-parents = 7
 max-attributes = 10
 min-public-methods = 2
